% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa_shp.R
\docType{data}
\name{africa_shp}
\alias{africa_shp}
\title{Africa shapefile data}
\format{
An \code{sf} object with spatial features.
}
\source{
Your data source
}
\usage{
africa_shp
}
\description{
A dataset containing spatial polygons of Africa.
}
\keyword{datasets}
