% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa_shps.R
\docType{data}
\name{africa_shps}
\alias{africa_shps}
\title{Africa shapefile data 2}
\format{
An \code{sf} object with spatial features.
}
\source{
Your data source
}
\usage{
africa_shps
}
\description{
A dataset containing spatial polygons of Africa.
}
\keyword{datasets}
