% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_autocorrelation.R
\name{compute_spatial_autocorr}
\alias{compute_spatial_autocorr}
\title{Compute Moran's I & LISA, classify clusters}
\usage{
compute_spatial_autocorr(sf_data, values, signif = 0.05)
}
\description{
Compute Moran's I & LISA, classify clusters
}
