% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{global_variables_eval}
\alias{global_variables_eval}
\title{Declare known global variables to suppress R CMD check NOTE
Global variables used in evaluation functions}
\description{
This is to suppress R CMD check notes about undefined global variables.
}
