% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{join_data}
\alias{join_data}
\title{Join spatial and incidence datasets}
\usage{
join_data(sf_data, tbl_data, by)
}
\arguments{
\item{sf_data}{sf object}

\item{tbl_data}{tibble of incidence}

\item{by}{Column name to join on}
}
\value{
sf object with joined attributes
}
\description{
Join spatial and incidence datasets
}
