% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_shapefile.R
\name{load_incidence_data}
\alias{load_incidence_data}
\title{Load incidence data from Excel}
\usage{
load_incidence_data(xlsx_path)
}
\arguments{
\item{xlsx_path}{Path to Excel file}
}
\value{
tibble of data
}
\description{
Load incidence data from Excel
}
