% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_shapefile.R
\name{load_shapefile}
\alias{load_shapefile}
\title{Load shapefile as sf + optionally convert to sp}
\usage{
load_shapefile(shp_path, to_sp = FALSE)
}
\arguments{
\item{shp_path}{Path to shapefile (.shp)}

\item{to_sp}{logical: also return Spatial object?}
}
\value{
list with sf and optionally sp object
}
\description{
Load shapefile as sf + optionally convert to sp
}
