% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_map_grid}
\alias{plot_map_grid}
\title{Arrange multiple tmap plots in a grid}
\usage{
plot_map_grid(maps, ncol = 2)
}
\arguments{
\item{maps}{List of tmap objects}

\item{ncol}{Number of columns}
}
\description{
Arrange multiple tmap plots in a grid
}
