% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{plot_obs_vs_pred}
\alias{plot_obs_vs_pred}
\title{Plot observed vs predicted}
\usage{
plot_obs_vs_pred(observed, predicted, title = "")
}
\arguments{
\item{observed}{A numeric vector of observed values}

\item{predicted}{A numeric vector of predicted values}

\item{title}{A string for the plot title}
}
\description{
Plot observed vs predicted
}
