% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_single_map}
\alias{plot_single_map}
\title{Build a tmap for a single variable}
\usage{
plot_single_map(sf_data, var, title, palette = "reds")
}
\arguments{
\item{sf_data}{sf object with variable}

\item{var}{variable name as string}

\item{title}{legend title}

\item{palette}{Color palette for map}
}
\value{
tmap object
}
\description{
Build a tmap for a single variable
}
