% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_MLmodels.R
\name{train_svr_examples}
\alias{train_svr_examples}
\title{SVR training examples}
\description{
SVR training examples
}
\examples{
svr_model <- train_svr(
  mapdata,
  incidence ~ female + male + agea + ageb + agec + fagea + fageb + fagec +
              magea + mageb + magec + yrb + yrc + yrd + yre
)
summary(svr_model)
}
