% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_MLmodels.R
\name{train_xgb_examples}
\alias{train_xgb_examples}
\title{XGBoost training examples}
\description{
XGBoost training examples
}
\examples{
xgb_model <- train_xgb(
  mapdata,
  incidence ~ female + male + agea + ageb + agec + fagea + fageb + fagec +
              magea + mageb + magec + yrb + yrc + yrd + yre,
  nrounds = 100
)
summary(xgb_model)
}
