% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{eval_model}
\alias{eval_model}
\title{Get RMSE/MAE/R² metrics on training data}
\usage{
eval_model(model, data, formula, model_type = c("rf", "xgb", "svr"))
}
\description{
Get RMSE/MAE/R² metrics on training data
}
