% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{eval_model}
\alias{eval_model}
\title{Get RMSE/MAE/R² metrics on training data}
\usage{
eval_model(model, data, formula, model_type = c("rf", "xgb", "svr"))
}
\arguments{
\item{model}{A trained model}

\item{data}{A data frame}

\item{formula}{A formula object}

\item{model_type}{Character string: one of "rf", "xgb", or "svr"}
}
\value{
A numeric value representing the model's accuracy
}
\description{
Evaluate Model Performance by calculating RMSE, MAE, and R² metrics.
}
