% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{model_evaluation_examples}
\alias{model_evaluation_examples}
\title{Examples for model evaluation functions}
\description{
Examples for model evaluation functions
}
\examples{
\dontrun{
library(caret)
data(iris)
rf_model <- train(Sepal.Length ~ ., data = iris, method = "rf")
eval_model(rf_model, iris, Sepal.Length ~ ., "rf")
}
}
