% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_map_grid}
\alias{plot_map_grid}
\title{Arrange Multiple tmap Plots in a Grid}
\usage{
plot_map_grid(maps, ncol = 2)
}
\arguments{
\item{maps}{A list of tmap objects.}

\item{ncol}{Number of columns in the grid (default is 2).}
}
\value{
A tmap object representing arranged maps.
}
\description{
Arrange a list of tmap objects into a grid layout.
}
\examples{
\dontrun{
# Assume 'mapdata' is an sf object with spatial data and variables
p1 <- plot_single_map(mapdata, "incidence", "Incidence")
p2 <- plot_single_map(mapdata, "female", "Female")
p3 <- plot_single_map(mapdata, "male", "Male")
plot_map_grid(list(p1, p2, p3), ncol = 3)
}
}
