% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_MLmodels.R
\name{train_svr}
\alias{train_svr}
\title{Train Support Vector Regression (SVR) model}
\usage{
train_svr(data, formula)
}
\arguments{
\item{data}{A data frame containing the training data.}

\item{formula}{A formula specifying the model.}
}
\value{
A trained \code{svm} model object from the \pkg{e1071} package.
}
\description{
Train Support
}
\details{
Trains an SVR model using the radial kernel.
}
\examples{
\dontrun{
svr_model <- train_svr(mapdata, incidence ~ female + male + agea + ageb + agec + fagea + fageb + fagec +
              magea + mageb + magec + yrb + yrc + yrd + yre)
summary(svr_model)
}
}
