% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_MLmodels.R
\name{train_xgb}
\alias{train_xgb}
\title{Train XGBoost model}
\usage{
train_xgb(data, formula, nrounds = 100, max_depth = 4, eta = 0.1)
}
\arguments{
\item{data}{A data frame with the training data.}

\item{formula}{A formula defining the model structure.}

\item{nrounds}{Number of boosting iterations.}

\item{max_depth}{Maximum tree depth.}

\item{eta}{Learning rate.}
}
\value{
A trained xgboost model object.
}
\description{
Train XGBoost model
}
\examples{
\dontrun{
rf_model <- train_rf(
  mapdata,
  incidence ~ female + male + agea + ageb + agec + fagea + fageb + fagec +
              magea + mageb + magec + yrb + yrc + yrd + yre,
  ntree = 500
)
}
Train XGBoost model

Trains an XGBoost regression model.

}
